+++ src/OpenTelemetry.Exporter.Geneva/GenevaExporterOptions.cs
@@ -1,8 +1,6 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-using System;
-using System.Collections.Generic;
 using System.Text;
 using OpenTelemetry.Internal;