+++ src/OpenTelemetry.Exporter.Geneva/Metrics/GenevaMetricExporter.cs
@@ -1,7 +1,6 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-using System;
 using System.Reflection;
 using System.Text.RegularExpressions;
 using OpenTelemetry.Exporter.Geneva.Metrics;