+++ src/OpenTelemetry.Exporter.Geneva/MsgPackExporter/MsgPackExporter.cs
@@ -1,11 +1,9 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-using System;
 #if NET8_0_OR_GREATER
 using System.Collections.Frozen;
 #endif
-using System.Collections.Generic;
 
 namespace OpenTelemetry.Exporter.Geneva;