+++ src/OpenTelemetry.Extensions.Enrichment/Internal/TraceEnrichmentActions.cs
@@ -1,10 +1,6 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
-
 namespace OpenTelemetry.Extensions.Enrichment;
 
 #pragma warning disable CA1812 // Class is instantiated through dependency injection