+++ test/OpenTelemetry.Sampler.AWS.Tests/TestRateLimiter.cs
@@ -1,10 +1,6 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-using System;
-using System.Collections.Generic;
-using System.Threading;
-using System.Threading.Tasks;
 using Xunit;
 
 namespace OpenTelemetry.Sampler.AWS.Tests;