+++ src/OpenTelemetry.Sampler.AWS/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+## 0.1.0-alpha.1
+
+Released 2024-Jun-20
+
 Initial release of `OpenTelemetry.Sampler.AWS`.
 
 * Feature - AWSXRayRemoteSampler - Add support for AWS X-Ray remote sampling
@@ -11,5 +15,5 @@ Initial release of `OpenTelemetry.Sampler.AWS`.
 * Make OpenTelemetry.Sampler.AWS native AoT compatible.
   ([#1541](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1541))
 
-* Update OpenTelemetry SDK version to `1.8.1`.
-  ([#1668](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1668))
+* Updated OpenTelemetry core component version(s) to `1.9.0`.
+  ([#1888](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1888))